relocates piece move logic to moves file
